Expanded property setting to include arrays, and sequences of arrays. -- also moved esprintf into a namespace (finally) - files with only OU::esprintf not mentioned -- several fixes to support workers with no ports at all components/Makefile: added ptest rcc component for property testing core/container/hdl_container/impl/src/HdlContainer.cxx: new property setting methods core/container/hdl_container/impl/src/XyzContainer.cxx: ditto core/container/interfaces/api/include/OcpiWorker.h:  ditto core/container/rcc_container/include/OcpiRccWorker.h: ditto core/container/rcc_container/src/OcpiRccWorker.cxx: ditto -- plus bug fix for workers WITH NO PORTS core/container/ocl_container/impl/src/OcpiOclContainer.cxx: ditto - but nonfunctional core/container/interfaces/api/src/OcpiWorker.cxx: ditto - and code changed to use them core/container/interfaces/api/include/OcpiContainerApi.h: better arg names core/local/util/assembly/src/OcpiUtilAssembly.cxx: add valuefile attribute -- this allows initial property values to come from files core/local/util/ezxml/include/OcpiUtilEzxml.h: remove esprintf from here core/local/util/ezxml/src/OcpiUtilEzxml.cxx: ditto core/local/util/misc/include/OcpiUtilMisc.h: add esprintf, plus fileString core/local/util/misc/src/OcpiUtilMisc.cxx: ditto tools/cdk/cdkutils/Makefile.ocpi.imports: add ocpios dependency tools/cdk/cdkutils/include/cdkutils.h: remove esprintf from here too tools/cdk/ocpigen/ocpigen/src/wip_HDL.cxx: properly support workers with no ports tools/cdk/ocpigen/ocpigen/src/wip_RCC.cxx: ditto components/specs/ptest_spec.xml: spec for new property tester components/ptest.rcc/Makefile: new ptest worker components/ptest.rcc/ptest.xml: ditto tools/cdk/examples/ptest/Makefile: ditto - the test app tools/cdk/examples/ptest/Makefile/ptest.xml

+++ core/local/util/property/src/OcpiUtilValue.cxx
@@ -41,6 +41,7 @@
 #include <string.h>
 #include "OcpiOsAssert.h"
 #include "OcpiUtilEzxml.h"
+#include "OcpiUtilMisc.h"
 #include "OcpiUtilAutoMutex.h"
 #include "OcpiUtilValue.h"
 #include "OcpiUtilDataTypes.h"
@@ -142,8 +143,8 @@ namespace OCPI {
 	  nBraces++; break;
 	case '}':
 	  if (nBraces == 0)
-	    return esprintf("unbalanced braces - extra close brace for (%*s) (%u)",
-			    stop - tmp, tmp, stop - tmp);
+	    return esprintf("unbalanced braces - extra close brace for (%*s) (%zu)",
+			    (int)(stop - tmp), tmp, stop - tmp);
 	  if (--nBraces == 0 && !comma) {
 	    end = unparsed;
 	    goto break2;
@@ -407,8 +408,8 @@ namespace OCPI {
 	  start++;
 	if (v->needsComma()) {
 	  if (*start != '{' || end[-1] != '}')
-	    return esprintf("struct member value needs be enclosed in {} (%*s) (%u)",
-			    end - start, start, end - start);
+	    return esprintf("struct member value needs be enclosed in {} (%*s) (%zu)",
+			    (int)(end - start), start, end - start);
 	  start++;
 	  end--;
 	}
@@ -602,8 +603,8 @@ namespace OCPI {
 	last = start;
 	if (nextDim < m_vt->m_arrayRank) {
 	  if (start == end || start[0] != '{' || end[-1] != '}')
-	    return esprintf("array elements not enclosed in {} for (%*s) (%u)",
-			    end - start, start, end - start);
+	    return esprintf("array elements not enclosed in {} for (%*s) (%zu)",
+			    (int)(end - start), start, end - start);
 	  if ((err = parseDimension(start+1, end-1, nseq, nextDim, offset, skip)))
 	    return err;
 	  offset += skip;
@@ -611,8 +612,8 @@ namespace OCPI {
 	  // strip braces if they were added
 	  if (needsCommaDimension()) {
 	    if (start[0] != '{' || end[-1] != '}')
-	      return esprintf("struct or array element not enclosed in {} for (%*s) (%u)",
-			      end - start, start, end - start);
+	      return esprintf("struct or array element not enclosed in {} for (%*s) (%zu)",
+			      (int)(end - start), start, end - start);
 	    start++;
 	    end--;
 	  }
@@ -640,7 +641,7 @@ namespace OCPI {
 				   m_vt->m_isSequence || m_vt->m_arrayRank ?      \
 				   m_p##pretty[nSeq * m_vt->m_nItems + nArray] : \
 				   m_##pretty)))			        \
-	    return esprintf("%s for \"%*s\" (%u)", err, end-start, start, end-start);\
+	    return esprintf("%s for \"%*s\" (%zu)", err, (int)(end-start), start, end-start); \
 	  break;
 	OCPI_PROPERTY_DATA_TYPES
 	OCPI_DATA_TYPE(sca,corba,letter,bits,StructValue,Struct,store)
